api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langgraph==0.2.3
langchain-openai==0.1.25
pydantic==2.10.3
sse-starlette==2.1.0
httpx==0.27.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
sqlalchemy-utils==0.41.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
firebase-admin==6.5.0
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.3
dj-database-url==2.1.0